### 2020-05-27 #2211 @arfon
Ahhh. @trallard - looks like you need to assign yourself first (this is a bug).
