### 2020-06-05 #2211 @trallard
Dear potential reviewers,

On behalf of the JOSS team I would like to ask if any of you's be able to review this JOSS submission:

**Title**:  seg1d: A Python package for Automated segmentation of one-dimensional (1D) data
**Repo**: https://github.com/cadop/seg1d
**Submitting author**: @cadop

@ixjlyons Kenneth, will you be willing to and are able to review this submission?
@alcantarar Ryan, will you be willing to and are able to review this submission?
@akuederle Arne,  will you be willing to and are able to review this submission?
@ lbugnon Leandro,   will you be willing to and are able to review this submission?

We are currently looking for at least one more reviewer to proceed.
