### 2020-06-22 9921a4f
Update paper.bib
Fixed DOIs
