+++ paper/paper.bib
@@ -17,14 +17,16 @@
   booktitle={Proceedings of the 2005 SIAM International Conference on Data Mining},
   pages={506--510},
   year={2005},
-  organization={SIAM}
+  organization={SIAM},
+  doi={https://doi.org/10.1137/1.9781611972757.50}
 }
 @inproceedings{kapadia2013efficient,
   title={Efficient motion retrieval in large motion databases},
   author={Kapadia, Mubbasir and Chiang, I-kao and Thomas, Tiju and Badler, Norman I and Kider Jr, Joseph T},
   booktitle={Proceedings of the ACM SIGGRAPH Symposium on Interactive 3D Graphics and Games},
   pages={19--28},
-  year={2013}
+  year={2013},
+  doi={https://doi.org/10.1145/2448196.2448199}
 }
 @inproceedings{bouchard2015segmenting,
   title={Segmenting motion capture data using a qualitative analysis},
@@ -32,7 +34,8 @@
   booktitle={Proceedings of the 8th ACM SIGGRAPH Conference on Motion in Games},
   pages={23--30},
   year={2015},
-  organization={ACM}
+  organization={ACM},
+  doi={https://doi.org/10.1145/2822013.2822039}
 }
 @article{agostini2013segmentation,
   title={Segmentation and classification of gait cycles},
@@ -42,7 +45,8 @@
   number={5},
   pages={946--952},
   year={2013},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={https://doi.org/10.1109/tnsre.2013.2291907}
 }
 @article{taborri2016gait,
   title={Gait partitioning methods: A systematic review},
@@ -52,7 +56,8 @@
   number={1},
   pages={66},
   year={2016},
-  publisher={Multidisciplinary Digital Publishing Institute}
+  publisher={Multidisciplinary Digital Publishing Institute},
+  doi={https://doi.org/10.3390/s16010066}
 }
 @inproceedings{jiang2017robust,
   title={A robust algorithm for gait cycle segmentation},
@@ -60,7 +65,8 @@
   booktitle={2017 25th European Signal Processing Conference (EUSIPCO)},
   pages={31--35},
   year={2017},
-  organization={IEEE}
+  organization={IEEE},
+  doi={https://doi.org/10.23919/eusipco.2017.8081163}
 }
 @article{sarsfield2019segmentation,
   title={Segmentation of Exercise Repetitions Enabling Real-Time Patient Analysis and Feedback Using a Single Exemplar},
@@ -70,21 +76,24 @@
   number={5},
   pages={1004--1019},
   year={2019},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={https://doi.org/10.1109/tnsre.2019.2907483}
 }
 @inproceedings{vlachos2003indexing,
   title={Indexing multi-dimensional time-series with support for multiple distance measures},
   author={Vlachos, Michail and Hadjieleftheriou, Marios and Gunopulos, Dimitrios and Keogh, Eamonn},
   booktitle={Proceedings of the ninth ACM SIGKDD international conference on Knowledge discovery and data mining},
   pages={216--225},
-  year={2003}
+  year={2003},
+  doi={https://doi.org/10.1145/956750.956777}
 }
 @inproceedings{keogh2004indexing,
   title={Indexing large human-motion databases},
   author={Keogh, Eamonn and Palpanas, Themistoklis and Zordan, Victor B and Gunopulos, Dimitrios and Cardle, Marc},
   booktitle={Proceedings of the Thirtieth international conference on Very large data bases-Volume 30},
   pages={780--791},
-  year={2004}
+  year={2004},
+  doi={https://doi.org/10.1016/b978-012088469-8.50069-3}
 }
[…]
