### 2020-06-29 #2404 @whedon
[REVIEW]: seg1d: A Python package for Automated segmentation of one-dimensional (1D) data
**Submitting author:** @cadop (<a href="http://orcid.org/0000-0003-3662-7203">Mathew Schwartz</a>)
**Repository:** <a href="https://github.com/cadop/seg1d" target ="_blank">https://github.com/cadop/seg1d</a>
**Version:** 0.1.0
