### 2020-06-29 cadop/seg1d#1 @AKuederle
[Joss Review] Author List
The paper contains multiple authors, but according to the git history only a single author contributed to the software project. I would be nice if the role and work of the other authors could be highlighted in the paper
