### 2020-06-29 #2404 @AKuederle
Started the review by opening issues in the repo:

https://github.com/cadop/seg1d/tree/master/paper
