### 2020-06-30 a57e72f
Remove egg and dist folder
Based on https://github.com/cadop/seg1d/issues/2
