+++ seg1d.egg-info/requires.txt
@@ -1,4 +1,0 @@
-numpy>=1.15
-scipy>=1.0.0
-scikit-learn>=0.2
-numba>=0.40.0
