+++ seg1d.egg-info/SOURCES.txt
@@ -1,28 +1,0 @@
-LICENSE.txt
-README.md
-setup.cfg
-setup.py
-seg1d/__init__.py
-seg1d/_about.py
-seg1d/algorithm.py
-seg1d/optimized_funcs.py
-seg1d/segment.py
-seg1d.egg-info/PKG-INFO
-seg1d.egg-info/SOURCES.txt
-seg1d.egg-info/dependency_links.txt
-seg1d.egg-info/requires.txt
-seg1d.egg-info/top_level.txt
-seg1d/examples/__init__.py
-seg1d/examples/ex_ecg.py
-seg1d/examples/ex_gauss.py
-seg1d/examples/ex_multi_signal.py
-seg1d/examples/ex_segment_class.py
-seg1d/examples/ex_segmenter_features.py
-seg1d/examples/ex_segmenter_sine.py
-seg1d/examples/ex_simple.py
-seg1d/examples/ex_sine.py
-seg1d/examples/ex_sine_noise.py
-seg1d/examples/ex_sine_noise_few.py
-seg1d/examples/ex_sine_scale.py
-seg1d/examples/noise.py
-seg1d/examples/test.py
