### 2020-06-30 fd79dac
Update README.md
Updated the alternatives section to further address https://github.com/cadop/seg1d/issues/3
