### 2020-06-30 cadop/seg1d#1 @cadop
Are you asking for the paper to include an author contribution section to be added? I am happy to include it if you would like, just from the JOSS readme I saw it says:

" JOSS papers are only expected to contain a limited set of metadata (see example below), a Statement of Need, Summary, Acknowledgements, and References sections. "

In any case, below are the contributions. I labeled commits that were done based on the feedback given on email. I list them below each co-author.

Todd

Provided insights to the need for segmentation processes from a clinical perspective.
Provided discussions on validity and alternatives to segmentation during creation of the algorithm.
Reviewed and tested the code, giving feedback on usability.
Reviewed and gave feedback on the paper.

a3ab154ee38e3f518636598dc589c83b9b3f2e91
66fa42a4460f4ad72166900d10aef97cf6bf832e
a0d3ea1cbfc17ed0f4a648700baa5a15fd60dd6c
c10afef78fd719987fff78ecf950f54e04696ac0
7039dfbe0916f40dec09123f10285c34a312fcdb

Jon

Provided insights to the need for segmentation processes from a clinical perspective.
Provided funding that supported me.
Provided the data used to come up with the algorithm.
Reviewed and gave feedback on the paper.

c10afef78fd719987fff78ecf950f54e04696ac0
34327f2783c9f2420c0fa69f1a670f09dad0efb6
