### 2020-06-30 62460a0
Update paper.bib
Removed http:// part of doi's.
