+++ paper/paper.bib
@@ -18,7 +18,7 @@
   pages={506--510},
   year={2005},
   organization={SIAM},
-  doi={https://doi.org/10.1137/1.9781611972757.50}
+  doi={10.1137/1.9781611972757.50}
 }
 @inproceedings{kapadia2013efficient,
   title={Efficient motion retrieval in large motion databases},
@@ -26,7 +26,7 @@
   booktitle={Proceedings of the ACM SIGGRAPH Symposium on Interactive 3D Graphics and Games},
   pages={19--28},
   year={2013},
-  doi={https://doi.org/10.1145/2448196.2448199}
+  doi={10.1145/2448196.2448199}
 }
 @inproceedings{bouchard2015segmenting,
   title={Segmenting motion capture data using a qualitative analysis},
@@ -35,7 +35,7 @@
   pages={23--30},
   year={2015},
   organization={ACM},
-  doi={https://doi.org/10.1145/2822013.2822039}
+  doi={10.1145/2822013.2822039}
 }
 @article{agostini2013segmentation,
   title={Segmentation and classification of gait cycles},
@@ -46,7 +46,7 @@
   pages={946--952},
   year={2013},
   publisher={IEEE},
-  doi={https://doi.org/10.1109/tnsre.2013.2291907}
+  doi={10.1109/tnsre.2013.2291907}
 }
 @article{taborri2016gait,
   title={Gait partitioning methods: A systematic review},
@@ -57,7 +57,7 @@
   pages={66},
   year={2016},
   publisher={Multidisciplinary Digital Publishing Institute},
-  doi={https://doi.org/10.3390/s16010066}
+  doi={10.3390/s16010066}
 }
 @inproceedings{jiang2017robust,
   title={A robust algorithm for gait cycle segmentation},
@@ -66,7 +66,7 @@
   pages={31--35},
   year={2017},
   organization={IEEE},
-  doi={https://doi.org/10.23919/eusipco.2017.8081163}
+  doi={10.23919/eusipco.2017.8081163}
 }
 @article{sarsfield2019segmentation,
   title={Segmentation of Exercise Repetitions Enabling Real-Time Patient Analysis and Feedback Using a Single Exemplar},
@@ -77,7 +77,7 @@
   pages={1004--1019},
   year={2019},
   publisher={IEEE},
-  doi={https://doi.org/10.1109/tnsre.2019.2907483}
+  doi={10.1109/tnsre.2019.2907483}
 }
 @inproceedings{vlachos2003indexing,
   title={Indexing multi-dimensional time-series with support for multiple distance measures},
@@ -85,7 +85,7 @@
   booktitle={Proceedings of the ninth ACM SIGKDD international conference on Knowledge discovery and data mining},
   pages={216--225},
   year={2003},
-  doi={https://doi.org/10.1145/956750.956777}
+  doi={10.1145/956750.956777}
 }
 @inproceedings{keogh2004indexing,
   title={Indexing large human-motion databases},
@@ -93,7 +93,7 @@
   booktitle={Proceedings of the Thirtieth international conference on Very large data bases-Volume 30},
   pages={780--791},
   year={2004},
-  doi={https://doi.org/10.1016/b978-012088469-8.50069-3}
+  doi={10.1016/b978-012088469-8.50069-3}
 }
 @inproceedings{zhu2002statstream,
   title={Statstream: Statistical monitoring of thousands of data streams in real time},
@@ -109,7 +109,7 @@
   booktitle={Proceedings of the eleventh ACM SIGKDD international conference on Knowledge discovery in data mining},
   pages={743--749},
   year={2005},
-  doi={https://dl.acm.org/doi/10.1145/1081870.1081966}
+  doi={10.1145/1081870.1081966}
 }
[…]
