### 2020-07-01 cadop/seg1d#1 @AKuederle
Thank you! I am just asking, because it is one of the points on the review checklist to evaluate if all authors contributed to the developed software.

@trallard How do you suggest we handle this in this case
