### 2020-07-07 86a1942
Update README.md
added numpydoc dependency and matplotlib
