### 2020-07-07 a85ebda
Update install.rst
Fixing requirements for documentation
