+++ docs/source/install.rst
@@ -8,7 +8,7 @@ Currently tested on ``Python 3.8+``.
 
 For the package: ``numpy>=1.18``, ``scipy>=1.4.1``, ``sklearn>=0.22``, ``numba>=0.48``
 
-For building documentation: ``sphinx``, ``scipy`` html_theme
+For building documentation: ``sphinx``, ``scipy`` html_theme, ``numpydoc``, ``matplotlib``
 
 Get seg1d
 -------------
