+++ docs/source/generated/seg1d.optimized_funcs.vcor.rst
@@ -1,6 +1,6 @@
 seg1d.optimized\_funcs.vcor
 ===========================
 
-.. currentmodule:: seg1d
+.. currentmodule:: seg1d.optimized_funcs
 
-.. autoattribute:: optimized_funcs.vcor
+.. autofunction:: vcor
