### 2020-07-07 4f19178
Remove numba func signature for documentation in conf.py
Some error on building for Windows was happening with a newer version of the libraries.  Removing the method seems to resolve it.
