### 2020-07-07 c86c48c
Update install.rst
Add some easier to follow pip command for getting documentation requirements.
