### 2020-07-07 63eaad9
Update docs
Updating to include documentation dependency install with pip
