+++ docs/build/doctest/output.txt
@@ -1,194 +1,0 @@
-Results of doctest builder run on 2020-05-04 19:29:53
-=====================================================
-
-Document: generated/seg1d.algorithm.cluster
--------------------------------------------
-1 items passed all tests:
-   8 tests in default
-8 tests in 1 items.
-8 passed and 0 failed.
-Test passed.
-
-Document: generated/seg1d.Segmenter.clear_reference
----------------------------------------------------
-1 items passed all tests:
-   7 tests in default
-7 tests in 1 items.
-7 passed and 0 failed.
-Test passed.
-
-Document: api_ecg
------------------
-1 items passed all tests:
-  25 tests in default
-25 tests in 1 items.
-25 passed and 0 failed.
-Test passed.
-
-Document: generated/seg1d.Segmenter.t_segments
-----------------------------------------------
-1 items passed all tests:
-  11 tests in default
-11 tests in 1 items.
-11 passed and 0 failed.
-Test passed.
-
-Document: generated/seg1d.optimized_funcs.vcor
-----------------------------------------------
-1 items passed all tests:
-   5 tests in default
-5 tests in 1 items.
-5 passed and 0 failed.
-Test passed.
-
-Document: generated/seg1d.Segmenter.add_reference
--------------------------------------------------
-1 items passed all tests:
-  10 tests in default
-10 tests in 1 items.
-10 passed and 0 failed.
-Test passed.
-
-Document: api_feat
-------------------
-1 items passed all tests:
-  17 tests in default
-17 tests in 1 items.
-17 passed and 0 failed.
-Test passed.
-
-Document: generated/seg1d.optimized_funcs.rcor
-----------------------------------------------
-1 items passed all tests:
-   5 tests in default
-5 tests in 1 items.
-5 passed and 0 failed.
-Test passed.
-
-Document: generated/seg1d.Segmenter.segment
--------------------------------------------
-1 items passed all tests:
-   5 tests in default
-5 tests in 1 items.
-5 passed and 0 failed.
-Test passed.
-
-Document: generated/seg1d.algorithm.combine_corr
-------------------------------------------------
-1 items passed all tests:
-  11 tests in default
-11 tests in 1 items.
-11 passed and 0 failed.
-Test passed.
-
-Document: start
----------------
-**********************************************************************
-File "start.rst", line 38, in default
-Failed example:
-    print(np.around(segments,decimals=7))
-Expected:
-    [[9.500000e+02 1.050000e+03 1.000000e+00]
-     [7.550000e+02 8.540000e+02 9.867665e-01]
-     [1.146000e+03 1.245000e+03 9.867665e-01]
-     [1.343000e+03 1.441000e+03 9.498135e-01]
-     [5.590000e+02 6.570000e+02 9.498135e-01]
-     [1.540000e+03 1.638000e+03 8.949109e-01]
-     [3.620000e+02 4.600000e+02 8.949109e-01]
-     [1.738000e+03 1.836000e+03 8.301899e-01]
-     [1.640000e+02 2.620000e+02 8.301899e-01]]
-Got:
-    [[9.500000e+02 1.050000e+03 1.000000e+00]
-     [1.146000e+03 1.245000e+03 9.867665e-01]
-     [7.550000e+02 8.540000e+02 9.867665e-01]
-     [1.343000e+03 1.441000e+03 9.498135e-01]
-     [5.590000e+02 6.570000e+02 9.498135e-01]
-     [1.540000e+03 1.638000e+03 8.949109e-01]
-     [3.620000e+02 4.600000e+02 8.949109e-01]
-     [1.738000e+03 1.836000e+03 8.301899e-01]
-     [1.640000e+02 2.620000e+02 8.301899e-01]]
-**********************************************************************
-1 items had failures:
[…]
