### 2020-07-08 0134881
Update setup.py
Directly import version so module is not run (which tries to import dependencies)
