### 2020-07-13 65f1464
Update python-package.yml
Trying to limit flake8 to seg1d directory
