### 2020-07-13 f08ff27
Update python-package.yml
Add documentation dependency
