+++ .github/workflows/python-package.yml
@@ -30,7 +30,7 @@ jobs:
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
     - name: Install documentation dependencies
       run: |
-        pip install sphinx numpydoc      
+        pip install sphinx numpydoc matplotlib     
     - name: Install package
       run: |
         pip install .
