### 2020-07-13 46aefc8
Update python-package.yml
Try to use internal test confirmation
