+++ .github/workflows/python-package.yml
@@ -40,6 +40,10 @@ jobs:
         flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics --filename=./seg1d/*.py
         # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Check code test works
+      run: |
+        from seg1d.examples import test
+        test.run()
     - name: Test with doctest
       run: |
         pytest --doctest-modules
