+++ .github/workflows/python-package.yml
@@ -42,9 +42,7 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Check code test works
       run: |
-        python3
-        from seg1d.examples import test
-        test.run()
+        python -c "from seg1d.examples import test; test.run()"
     - name: Test with doctest
       run: |
         pytest --doctest-modules
