+++ .github/workflows/python-package.yml
@@ -33,7 +33,7 @@ jobs:
         pip install sphinx numpydoc matplotlib     
     - name: Install package
       run: |
-        pip install .
+        pip install -e .
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
