### 2020-07-13 cad2e23
Update ex_sine_noise_few.py
fix test case
