### 2020-07-16 #2404 @trallard
Hi folks just checking on the review.

@ejhigson do you have any updates on the review here? I know these are complicated times and your contribution and time are appreciated.
