### 2020-07-19 #2404 @ejhigson
@cadop congrations on a nice software package and paper! I have installed it and successfully run the tests and examples, as well as reviewing the documentation and paper. I only have a few minor questions:

1) I see the minimum versions for the dependencies in ``setup.py`` are different to those which are listed in the documentation - should this be updated?

2) I get an error running the code in the README as I think you need to change ``seg1d.segmentData`` to ``seg1d.segment_data`` (as in the documentation).

3) I see that there are three authors on the paper, but that you appear to be the only code contributor to the github module. Please can you confirm the role of the other two authors in the software package, and that they meet the [JOSS authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship)?
