### 2020-07-19 #2404 @AKuederle
To provide a summary of my feedback:

First of all, the package is very well written and documented. It fullfils the purpose, which is very well outlined in the paper, nicely and thanks to the examples it should be easy to get it running on a new dataset.

From a scientific perspective, this package mainly wraps a template matching approach based on cross correlation with a subsequent clustering with a nice interface. While, in particular the template matching part could be considered a very common approach to my (and the authors) knowledge, there is no simple implementation out there. Therefore, a nice implementation with any easy-to-use interface is definitely valuable.

However, I during the review, I had a couple of concerns regarding the overall amount of content in the software package:

1. The name `seg1d` suggests that this is a package that implements a set of possible ways to segment a time series, but it just implements a single method.
2. Second, given that it just implements a single method, from a user perspective, I would burden myself with an additional dependency. My suggestion (independent of this paper and review) to author would be, to try to get a modified version of the implementation integrated into one of the already popular timeseries analysis package. This would ensure that people can find and use the method easily, and would increase the chance that the software is properly maintained for a longer time.

Besides the points above, I would fully recommend the package for publication. It is a nice piece of software with a clear usecase.
