### 2020-07-20 d01632e
Update README.md
fixed example function name
