### 2020-07-20 dc007ac
Update README.md
expanded required versions for dependencies
