+++ README.md
@@ -38,7 +38,7 @@ Currently tested on ``Python 3.8`` on Ubuntu 18.04 and Windows 10. (Should work 
 
 Required Packages:
 
-``numpy>=1.18``, ``scipy>=1.4.1``, ``sklearn>=0.22``, ``numba>=0.48``
+``numpy>=1.15``, ``scipy>=1.0.0``, ``sklearn>=0.2``, ``numba>=0.40``
 
 For documentation:
 
