### 2020-07-20 efe7358
Update install.rst
expanded required dependencies version
