### 2020-07-20 #2404 @cadop
@ejhigson  Thank you for pointing out the readme error!  The commits below correspond to your items.

1) https://github.com/cadop/seg1d/commit/dc007ac4115de3df8e555278d322b33f1f41fafc https://github.com/cadop/seg1d/commit/efe73589648c21f4b1e1fa4c309ae77cd87e746d

2) https://github.com/cadop/seg1d/commit/d01632e3d497aa098b1a31b09e3d7abf2e65da45

3) https://github.com/cadop/seg1d/issues/1
