### 2020-07-22 37b669e
Update README.md
add reference to published proceedings
