### 2020-07-23 bc69a5c
Update python-package.yml
Add a final check on documentation
