### 2020-07-25 #2404 @ejhigson
Hi @cadop - thank you very much for these changes, everything looks good to me! I am now happy to recommend the paper for publication in JOSS. Congratulations on a nice software package!
