### 2020-07-27 #2404 @cadop
@ejhigson Thank you very much!

@AKuederle Thank you for the in-depth review and time you spent on considering the package for the community.  Please let me know if we need to continue some of the issue discussions as I have left them for you to close in your perspective.  As for the points you raise, I think its overall a fair assessment.  For some clarity, the package name is similar to `mwarp1d` due to an initial goal of combining them in a sort of ecosystem.

@trallard  I am not sure you received the notification, but you were tagged in this issue https://github.com/cadop/seg1d/issues/1 and I believe it is just waiting for you to close it.

Please let me know if there are any more steps/actions needed by me.
