### 2020-08-02 f9c94f3
add asserts to check for types
Probably not a good way to do this and instead index it as a numpy method to make it fail instead. assert is too slow.
