### 2020-08-11 cadop/seg1d#1 @trallard
Hey folks - I have checked our docs and processes and it should be ok to refer to acceptance as is. Though we do require that the authors have contributed to the software, @cadop  has provided good evidence that they all made significant contributions towards the development of the software. So if everyone is ok with this you can close the issue and I will move the review forward.
