### 2020-08-11 #2404 @trallard
@cadop I have created a [PR](https://github.com/cadop/seg1d/pull/10) to make some minor edits to the paper for acceptance.

@AKuederle if you're happy with the authoring resolution can you complete your checklist, please?
