### 2020-08-11 cadop/seg1d#1 @AKuederle
If this is fine from your side, I will close the issue
