### 2020-08-11 #2404 @trallard
@AKuederle  and @ejhigson  thank you very much for your contribution as reviewers for JOSS. Your time and contributions are much appreciated 🙌🏼
