### 2020-08-11 #2404 @trallard
@cadop since both reviewers have completed their duties and not flagged any major issues we are ready to move forward.

At this point could you:
