### 2020-08-11 #2404 @cadop
@trallard  Thanks for moving all these things along.

I made a release of the software https://github.com/cadop/seg1d/releases/tag/v0.1.2

and made a zenodo upload:
https://zenodo.org/record/3979649#.XzLab55KiUk

DOI 10.5281/zenodo.3979649

Is this all correct?

Thanks
