### 2020-08-17 #2404 @trallard
@cadop Thanks all seems in order. I have proceeded and completed my editorial duties. The Editor in Chief will take from here for acceptance.

Thanks everyone for your patience in these busy times and your contribution is much appreciated.
